FIX: 마이페이지 데이터가공 로직 위치 변경
